LayoutTests/platform/ios/imported/w3c/web-platform-tests/dom/nodes/moveBefore/selection-preserve-expected.txt

Four
One
Two
Three

FAIL moveBefore should not reset selection with preceding text assert_equals: expected Text node "This text does not move" but got null
FAIL moveBefore resets selection that enters a subtree, when the whole selection is moved assert_equals: expected Text node "Grandparent paragraph" but got null
FAIL moveBefore anchor node moved up to expand selection and absorb nodes assert_equals: expected Text node "Child paragraph one" but got null
FAIL moveBefore move intersecting nodes out of a selection assert_equals: expected Text node "Grandparent paragraph" but got null
FAIL moveBefore focus node moved up to shrink selection and exclude nodes; focus node gets reset assert_equals: expected Text node "Grandparent paragraph" but got null
FAIL moveBefore selection is not preserved, especially when underlying range gets inverted assert_equals: expected Text node "Three" but got null
